+++ test/logging.spec.js
@@ -1,0 +1,12 @@
+/* eslint-env mocha */
+
+'use strict'
+
+/**
+ * The code under test.
+ * @type {any}
+ */
+const T = require('../index.js')
+
+describe('the logging functionality', () => {
+})