+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+* All output from the spawned child's `stdout` and `stderr` can be saved to a log file. The path to this file will be returned by `stop()`.
+
+### Changed
+
+* The `running` property is now called `isRunning`.
+
+
 ## [0.3.0] - 2021-03-18
 
 ### Added
@@ -15,6 +26,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 * The `stop()` method no longer resolves to the exit code of the child process. I was not able to get this to work, and decided it was not a significant feature. 
 
+
 ## [0.2.0] - 2021-03-17
 
 ### Changed
@@ -25,6 +37,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 
 * When checking the `stdout` of the child process for the expected characters, the presence of blank lines was causing a problem in some situations. This is now fixed.
+
 
 ## [0.1.0] - 2021-03-15